src/scrap_code/requirements.txt
# LangGraph and LangChain dependencies
langgraph==0.1.1
langchain==0.2.11
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
langchain-core>=0.2.0

# OpenAI and embedding dependencies
openai>=1.0.0
chromadb>=0.5.0

# Environment and utility dependencies
python-dotenv>=1.0.0
typing-extensions>=4.0.0 

# WhatsApp Message Fetcher dependencies
fastapi>=0.100.0
uvicorn==0.23.2
pydantic>=2.0.0
requests>=2.31.0
aiohttp>=3.8.0
psycopg2-binary>=2.9.9

pytest
pydub>=0.25.0  # For audio format conversion